Move HTTPClient+BetaTester to BetaTester folder for consistency
